fix: Clean up code
